feat(routing): candidate universe as inspectable decision evidence
Every ordered routing decision records where its candidate enumeration
came from (universe_source, honestly CONFIGURED today) and what the
governed catalogue holds that could not serve - MODEL_NOT_CATALOGUED,
LIFECYCLE_INELIGIBLE, or POLICY_EXCLUDED, the operator story
configuration cannot tell (issue #244, U1). The enumeration is untouched;
derive_candidate_universe is proven equivalent to the configured pair by
test, so the U2 flip to catalogue-derived enumeration is visible
evidence, never a silent rewire.

The module-budget ratchet fired and got real splits, not raised
ceilings: contracts/provider_catalog.py takes the catalog/policy/profile
read-model family out of contracts/providers.py, and
services/model_lifecycle_control.py takes the #245 governed state
changes out of services/model_catalogue.py. Pure moves, import sites
updated, no compatibility shims.
